# backend_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mine Worker Safety Management System API",
    "description": "Backend API for Role-Based Access Control, Shift Monitoring, Hazard Reporting, Live Tracking, and SOS Alerts.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/locations/": {
      "post": {
        "tags": [
          "Locations"
        ],
        "summary": "Update Location",
        "operationId": "update_location_api_locations__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/locations/history/{worker_id}": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get Location History",
        "operationId": "get_location_history_api_locations_history__worker_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationOut"
                  },
                  "title": "Response Get Location History Api Locations History  Worker Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hazards/report": {
      "post": {
        "tags": [
          "Hazards"
        ],
        "summary": "Report Hazard",
        "operationId": "report_hazard_api_hazards_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HazardReportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/hazards/{hazard_id}/upload": {
      "post": {
        "tags": [
          "Hazards"
        ],
        "summary": "Upload Hazard Image",
        "operationId": "upload_hazard_image_api_hazards__hazard_id__upload_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hazard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Hazard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_hazard_image_api_hazards__hazard_id__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hazards/": {
      "get": {
        "tags": [
          "Hazards"
        ],
        "summary": "List Hazards",
        "operationId": "list_hazards_api_hazards__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HazardReportOut"
                  },
                  "type": "array",
                  "title": "Response List Hazards Api Hazards  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/hazards/{hazard_id}": {
      "get": {
        "tags": [
          "Hazards"
        ],
        "summary": "Get Hazard",
        "operationId": "get_hazard_api_hazards__hazard_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hazard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Hazard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hazards/{hazard_id}/assign": {
      "put": {
        "tags": [
          "Hazards"
        ],
        "summary": "Assign Investigation",
        "operationId": "assign_investigation_api_hazards__hazard_id__assign_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hazard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Hazard Id"
            }
          },
          {
            "name": "investigator_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investigator Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hazards/{hazard_id}/resolve": {
      "put": {
        "tags": [
          "Hazards"
        ],
        "summary": "Resolve Hazard",
        "operationId": "resolve_hazard_api_hazards__hazard_id__resolve_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hazard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Hazard Id"
            }
          },
          {
            "name": "remarks",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Remarks"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_hazard_image_api_hazards__hazard_id__upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_hazard_image_api_hazards__hazard_id__upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HazardImageOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_url",
          "created_at"
        ],
        "title": "HazardImageOut"
      },
      "HazardReportCreate": {
        "properties": {
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "location": {
            "type": "string",
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "hazard_type",
          "severity",
          "description",
          "location"
        ],
        "title": "HazardReportCreate"
      },
      "HazardReportOut": {
        "properties": {
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "reporter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "investigator_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigator Id"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/HazardImageOut"
            },
            "type": "array",
            "title": "Images",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "hazard_type",
          "severity",
          "description",
          "location",
          "id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "HazardReportOut"
      },
      "LocationCreate": {
        "properties": {
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "LocationCreate"
      },
      "LocationOut": {
        "properties": {
          "latitude": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Latitude"
          },
          "longitude": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Longitude"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "worker_id": {
            "type": "integer",
            "title": "Worker Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "id",
          "worker_id",
          "timestamp"
        ],
        "title": "LocationOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}